+++ src/utility/string_manipulation.hpp
@@ -150,7 +150,7 @@ std::string convert_fp_to_string(std::string num, unsigned long long precision);
 /**
  * convert a real number stored in a string into bits with a given precision
  */
-unsigned long long convert_fp_to_bits(std::string num, unsigned int precision);
+unsigned long long convert_fp_to_bits(std::string num, unsigned long long precision);
 
 /**
  * Macro returning the actual type of an object
